+++ ExpandedStorage/Constants.cs
@@ -2,11 +2,11 @@ namespace LeFauxMods.ExpandedStorage;
 
 internal static class Constants
 {
+    public const string BigCraftableData = "Data/BigCraftables";
+
     public const string ChestOpenSound = "openChest";
 
     public const string ChestPlacementSound = "axe";
-
-    public const string DataPath = "Data/BigCraftables";
 
     public const string LidCloseSound = "doorCreakReverse";